backend/requirements.txt
args==0.1.0
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
Automat==20.2.0
autopep8==1.5.4
awscli==1.18.183
backports.csv==1.0.7
beautifulsoup4==4.9.3
botocore==1.19.23
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
clint==0.5.1
colorama==0.4.3
constantly==15.1.0
contextlib2==0.5.5
cryptography==3.2.1
cssselect==1.1.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.1
django-admin-rangefilter==0.6.3
django-advanced-filters==1.2.0
django-braces==1.13.0
django-cors-headers==3.5.0
django-import-export==2.4.0
django-multiselectfield==0.1.12
django-mysql==3.9.0
django-searchable-select==1.5.0
djangorestframework==3.12.2
docopt==0.6.2
docutils==0.16
et-xmlfile==1.0.1
Flask==1.1.2
Flask-API==2.0
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
internetarchive==1.9.6
isort==5.6.4
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
jsonpatch==1.26
jsonpointer==2.0
lazy-object-proxy==1.4.3
lxml==4.6.1
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.1
odfpy==1.4.1
openpyxl==3.0.5
parsel==1.6.0
Pillow==8.0.1
Protego==0.1.16
psutil==5.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
PyMySQL==0.10.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
queuelib==1.5.0
requests==2.25.0
rsa==4.6
s3transfer==0.3.3
schema==0.7.2
Scrapy==2.4.1
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==2.0.0
toml==0.10.2
tqdm==4.53.0
Twisted==20.3.0
typed-ast==1.4.1
urllib3==1.26.2
w3lib==1.22.0
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.2.0
